# docs/21-api-openapi.yaml
openapi: 3.1.0
info:
  title: Gym Management System API
  version: 1.0.0
  description: Machine-readable V1 contract. Base monetary unit is integer minor units (100 = 1 EGP). Arabic RTL owner UI.
servers:
- url: /api/v1
security:
- cookieAuth: []
...
paths:
...
  /members:
    get:
      tags:
      - Members
      parameters:
      - name: search
        in: query
        schema:
          type: string
      - name: archived
        in: query
        schema:
          type: boolean
          default: false
      - name: subscriptionState
        in: query
        schema:
          type: string
          enum:
          - NONE
          - SCHEDULED
          - ACTIVE
          - EXPIRED
      - name: hasDebt
        in: query
        schema:
          type: boolean
      - $ref: '#/components/schemas/Page'
      - $ref: '#/components/schemas/Limit'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - Members
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemberRequest'
      responses:
        '201':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /members/{memberId}:
    get:
      tags:
      - Members
      parameters:
      - $ref: '#/components/schemas/MemberId'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - Members
      parameters:
      - $ref: '#/components/schemas/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRequest'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /members/{memberId}/archive:
    post:
      tags:
      - Members
      parameters:
      - $ref: '#/components/schemas/MemberId'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /members/{memberId}/restore:
    post:
      tags:
      - Members
      parameters:
      - $ref: '#/components/schemas/MemberId'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /members/{memberId}/photo:
    post:
      tags:
      - Members
      parameters:
      - $ref: '#/components/schemas/MemberId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoResponse'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '413':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags:
      - Members
      parameters:
      - $ref: '#/components/schemas/MemberId'
      responses:
        '200':
          description: Normalized member photo
          content:
            image/webp:
              schema:
                type: string
                format: binary
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /members/{memberId}/subscriptions:
    get:
      tags:
      - Subscriptions
      parameters:
      - $ref: '#/components/schemas/MemberId'
      - $ref: '#/components/schemas/Page'
      - $ref: '#/components/schemas/Limit'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - Subscriptions
      parameters:
      - $ref: '#/components/schemas/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionCreateResponse'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /members/{memberId}/payments:
    get:
      tags:
      - Payments
      parameters:
      - $ref: '#/components/schemas/MemberId'
      - $ref: '#/components/schemas/Page'
      - $ref: '#/components/schemas/Limit'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - Payments
      parameters:
      - $ref: '#/components/schemas/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /members/{memberId}/audit-log:
    get:
      tags:
      - Audit
      parameters:
      - $ref: '#/components/schemas/MemberId'
      - $ref: '#/components/schemas/Page'
      - $ref: '#/components/schemas/Limit'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditListResponse'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /audit-logs:
    get:
      tags:
      - Audit
      parameters:
      - name: entityType
        in: query
        schema:
          type: string
          enum:
          - MEMBER
          - PLAN
          - SUBSCRIPTION
          - PAYMENT
      - name: action
        in: query
        schema:
          type: string
          enum:
          - MEMBER_CREATED
          - MEMBER_UPDATED
          - MEMBER_ARCHIVED
          - MEMBER_RESTORED
          - PLAN_CREATED
          - PLAN_UPDATED
          - PLAN_ENABLED
          - PLAN_DISABLED
          - SUBSCRIPTION_CREATED
          - SUBSCRIPTION_UPDATED
          - SUBSCRIPTION_VOIDED
          - PAYMENT_CREATED
          - PAYMENT_VOIDED
      - name: entityId
        in: query
        schema:
          type: string
      - name: from
        in: query
        schema:
          type: string
          format: date
      - name: to
        in: query
        schema:
          type: string
          format: date
      - $ref: '#/components/schemas/Page'
      - $ref: '#/components/schemas/Limit'
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditListResponse'
        '400':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: خطأ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gym_session
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required:
      - statusCode
      - code
      - message
      - details
      properties:
        statusCode:
          type: integer
        code:
          type: string
        message:
          type: string
        details:
          anyOf:
          - type: object
            additionalProperties: true
          - type: 'null'
    Pagination:
      type: object
      additionalProperties: false
      required:
      - page
      - limit
      - totalItems
      - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    Subscription:
      type: object
      additionalProperties: false
      required:
      - id
      - memberId
      - planId
      - planNameSnapshot
      - durationMonths
      - listedPriceMinor
      - agreedPriceMinor
      - startDate
      - endDate
      - state
      - voidedAt
      - voidReason
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        planNameSnapshot:
          type: string
        durationMonths:
          type: integer
          enum:
          - 1
          - 3
          - 6
          - 12
        listedPriceMinor:
          type: integer
          minimum: 0
        agreedPriceMinor:
          type: integer
          minimum: 0
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        state:
          type: string
          enum:
          - SCHEDULED
          - ACTIVE
          - EXPIRED
          - VOIDED
        voidedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        voidReason:
          anyOf:
          - type: string
          - type: 'null'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Payment:
      type: object
      additionalProperties: false
      required:
      - id
      - memberId
      - amountMinor
      - paymentDate
      - paymentMethod
      - receiptNumber
      - balanceAfterPaymentMinor
      - status
      - voidedAt
      - voidReason
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        amountMinor:
          type: integer
          minimum: 1
        paymentDate:
          type: string
          format: date
        paymentMethod:
          type: string
          const: CASH
        receiptNumber:
          type: string
        balanceAfterPaymentMinor:
          type: integer
          minimum: 0
        status:
          type: string
          enum:
          - VALID
          - VOIDED
        voidedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        voidReason:
          anyOf:
          - type: string
          - type: 'null'
        createdAt:
          type: string
          format: date-time
    MemberSummary:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - phone
      - photoUrl
      - isArchived
      - subscriptionState
      - subscriptionEndDate
      - outstandingBalanceMinor
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        photoUrl:
          anyOf:
          - type: string
          - type: 'null'
        isArchived:
          type: boolean
        subscriptionState:
          type: string
          enum:
          - NONE
          - SCHEDULED
          - ACTIVE
          - EXPIRED
        subscriptionEndDate:
          anyOf:
          - type: string
            format: date
          - type: 'null'
        outstandingBalanceMinor:
          type: integer
          minimum: 0
    MemberDetail:
      type: object
      additionalProperties: false
      required:
      - id
      - name
      - phone
      - gender
      - dateOfBirth
      - photoUrl
      - heightCm
      - weightKg
      - joinDate
      - isArchived
      - archivedAt
      - currentSubscription
      - nextSubscription
      - outstandingBalanceMinor
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        dateOfBirth:
          type: string
          format: date
        photoUrl:
          anyOf:
          - type: string
          - type: 'null'
        heightCm:
          anyOf:
          - type: number
            exclusiveMinimum: 0
            maximum: 300
          - type: 'null'
        weightKg:
          anyOf:
          - type: number
            exclusiveMinimum: 0
            maximum: 500
          - type: 'null'
        joinDate:
          type: string
          format: date
        isArchived:
          type: boolean
        archivedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        currentSubscription:
          anyOf:
          - $ref: '#/components/schemas/Subscription'
          - type: 'null'
        nextSubscription:
          anyOf:
          - $ref: '#/components/schemas/Subscription'
          - type: 'null'
        outstandingBalanceMinor:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuditLogItem:
      type: object
      additionalProperties: false
      required:
      - id
      - actorOwnerId
      - entityType
      - entityId
      - action
      - before
      - after
      - metadata
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        actorOwnerId:
          type: string
          format: uuid
        entityType:
          type: string
          enum:
          - MEMBER
          - PLAN
          - SUBSCRIPTION
          - PAYMENT
        entityId:
          type: string
        action:
          type: string
          enum:
          - MEMBER_CREATED
          - MEMBER_UPDATED
          - MEMBER_ARCHIVED
          - MEMBER_RESTORED
          - PLAN_CREATED
          - PLAN_UPDATED
          - PLAN_ENABLED
          - PLAN_DISABLED
          - SUBSCRIPTION_CREATED
          - SUBSCRIPTION_UPDATED
          - SUBSCRIPTION_VOIDED
          - PAYMENT_CREATED
          - PAYMENT_VOIDED
        before:
          anyOf:
          - type: object
            additionalProperties: true
          - type: 'null'
        after:
          anyOf:
          - type: object
            additionalProperties: true
          - type: 'null'
        metadata:
          anyOf:
          - type: object
            additionalProperties: true
          - type: 'null'
        createdAt:
          type: string
          format: date-time
...
    CreateMemberRequest:
      type: object
      additionalProperties: false
      required:
      - name
      - phone
      - gender
      - dateOfBirth
      - joinDate
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 150
        phone:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        dateOfBirth:
          type: string
          format: date
        heightCm:
          anyOf:
          - type: number
            exclusiveMinimum: 0
            maximum: 300
          - type: 'null'
        weightKg:
          anyOf:
          - type: number
            exclusiveMinimum: 0
            maximum: 500
          - type: 'null'
        joinDate:
          type: string
          format: date
    UpdateMemberRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 150
        phone:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        dateOfBirth:
          type: string
          format: date
        heightCm:
          anyOf:
          - type: number
            exclusiveMinimum: 0
            maximum: 300
          - type: 'null'
        weightKg:
          anyOf:
          - type: number
            exclusiveMinimum: 0
            maximum: 500
          - type: 'null'
        joinDate:
          type: string
          format: date
...
    InitialPaymentInput:
      type: object
      additionalProperties: false
      required:
      - amountMinor
      - paymentDate
      properties:
        amountMinor:
          type: integer
          minimum: 1
        paymentDate:
          type: string
          format: date
    CreateSubscriptionRequest:
      type: object
      additionalProperties: false
      required:
      - planId
      - durationMonths
      - startDate
      - agreedPriceMinor
      properties:
        planId:
          type: string
          format: uuid
        durationMonths:
          type: integer
          enum:
          - 1
          - 3
          - 6
          - 12
        startDate:
          type: string
          format: date
        agreedPriceMinor:
          type: integer
          minimum: 0
        initialPayment:
          anyOf:
          - $ref: '#/components/schemas/InitialPaymentInput'
          - type: 'null'
...
    CreatePaymentRequest:
      type: object
      additionalProperties: false
      required:
      - amountMinor
      - paymentDate
      properties:
        amountMinor:
          type: integer
          minimum: 1
        paymentDate:
          type: string
          format: date
...
    MemberListResponse:
      type: object
      additionalProperties: false
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MemberSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    SubscriptionListResponse:
      type: object
      additionalProperties: false
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Subscription'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PaymentListResponse:
      type: object
      additionalProperties: false
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        pagination:
          $ref: '#/components/schemas/Pagination'
    AuditListResponse:
      type: object
      additionalProperties: false
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogItem'
        pagination:
          $ref: '#/components/schemas/Pagination'
    SubscriptionCreateResponse:
      type: object
      additionalProperties: false
      required:
      - subscription
      - initialPayment
      - outstandingBalanceMinor
      properties:
        subscription:
          $ref: '#/components/schemas/Subscription'
        initialPayment:
          anyOf:
          - $ref: '#/components/schemas/Payment'
          - type: 'null'
        outstandingBalanceMinor:
          type: integer
          minimum: 0
...
    PhotoResponse:
      type: object
      additionalProperties: false
      required:
      - photoUrl
      properties:
        photoUrl:
          type: string
...